quantizers cleanup
